Official: "Why is Git like this" Update
the pain and suffering brought to us in this push was ridiculous. The player would not move. As a quick fix, I uncommented some code that had magically commented itself out and had to add time consideration into the movement. Quite surprisingly, the playercontroller seemed to be a hybrid of old and new.